pyproject.toml
[build-system]
requires = ["hatchling>=1.24"]
build-backend = "hatchling.build"

[project]
name = "ai-trading-system"
version = "0.1.0"
description = "AI sector trend analysis and position management system."
readme = "README.md"
requires-python = ">=3.11"
license = { text = "Proprietary" }
authors = [{ name = "wakare", email = "wakare_no_kaze@outlook.com" }]
dependencies = [
  "numpy>=1.26",
  "pandas>=2.2",
  "pydantic>=2.7",
  "pydantic-settings>=2.2",
  "PyYAML>=6.0",
  "rich>=13.7",
  "typer>=0.12",
]

[project.optional-dependencies]
data = [
  "pandas-datareader>=0.10",
  "yfinance>=0.2.40",
]
dashboard = [
  "plotly>=5.22",
  "streamlit>=1.36",
]
dev = [
  "mypy>=1.10",
  "pytest>=8.2",
  "ruff>=0.5",
  "types-PyYAML>=6.0.12",
]

[project.scripts]
aits = "ai_trading_system.cli:app"

[tool.hatch.build.targets.wheel]
packages = ["src/ai_trading_system"]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["src"]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "UP", "B"]
ignore = []

[tool.mypy]
python_version = "3.11"
strict = true
warn_unreachable = true
warn_unused_configs = true
